+++ tests/test_your_project.py
@@ -1,5 +1,5 @@
 """Test main cli"""
-import your_project.cli as cli
+import aletheia.cli as cli
 
 def test_get_message():
     assert cli.yo() == 'Yo cli!', 'cli.yo() returns "Yo cli!"'
